+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2.2.0
+  * Exclude unauthorized streams (403) from catalog during discovery [#29](https://github.com/singer-io/tap-toggl/pull/29)
+
 ## 2.1.0
   * Bump dependency versions for twistlock compliance [#37](https://github.com/singer-io/tap-toggl/pull/37)